+++ src/middlewares/cors.ts
@@ -4,7 +4,7 @@
 
 import { FritterContext } from "../classes/FritterContext.js";
 
-import type { MiddlewareFunction } from "../types/FritterMiddlewareFunction.js";
+import { MiddlewareFunction } from "../types/MiddlewareFunction.js";
 
 //
 // Interfaces